+++ mini-spring/src/main/java/beandefinitionandbeandefinitionregisty/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -1,0 +1,32 @@
+package beandefinitionandbeandefinitionregisty.beans.factory.support;
+
+import beandefinitionandbeandefinitionregisty.beans.BeansException;
+import beandefinitionandbeandefinitionregisty.beans.factory.config.BeanDefinition;
+
+/**
+ * @Author huabin
+ * @DateTime 2022-11-30 09:57
+ * @Desc
+ */
+public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory{
+
+
+    @Override
+    protected Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException {
+        return doCreateBean(beanName, beanDefinition);
+    }
+
+    protected Object doCreateBean(String beanName, BeanDefinition beanDefinition) {
+        Class beanClass = beanDefinition.getBeanClass();
+
+        Object bean = null;
+        try {
+            bean = beanClass.newInstance();
+        } catch (Exception e) {
+            throw new BeansException("Instantiation of bean failed", e);
+        }
+
+        addSingleton(beanName, bean);
+        return bean;
+    }
+}