+++ mini-spring/src/main/java/beandefinitionandbeandefinitionregisty/beans/factory/support/BeanDefinitionRegistry.java
@@ -1,0 +1,18 @@
+package beandefinitionandbeandefinitionregisty.beans.factory.support;
+
+import beandefinitionandbeandefinitionregisty.beans.factory.config.BeanDefinition;
+
+
+/**
+ * BeanDefinition注册表接口
+ */
+public interface BeanDefinitionRegistry {
+
+    /**
+     * 向注册表中注册BeanDefinition
+     * @param beanName
+     * @param beanDefinition
+     */
+    void registerBeanDefinition(String beanName, BeanDefinition beanDefinition);
+
+}