+++ mini-spring/changelog.md
@@ -6,3 +6,9 @@
 ## BeanDefinition和BeanDefinitionRegistry
 > package:beandefinitionandbeandefinitionregisty
 
+主要增加如下类：
+
+- BeanDefinition，顾名思义，用于定义bean信息的类，包含bean的class类型、构造参数、属性值等信息，每个bean对应一个BeanDefinition的实例。简化BeanDefinition仅包含bean的class类型。
+- BeanDefinitionRegistry，BeanDefinition注册表接口，定义注册BeanDefinition的方法。
+- SingletonBeanRegistry及其实现类DefaultSingletonBeanRegistry，定义添加和获取单例bean的方法。
+bean容器作为BeanDefinitionRegistry和SingletonBeanRegistry的实现类，具备两者的能力。向bean容器中注册BeanDefinition后，使用bean时才会实例化。